Fix edit employee
